+++ moPepGen/circ/CircRNA.py
@@ -30,7 +30,7 @@ class CircRNAModel():
         self.transcript_id = transcript_id
         self.gene_name = gene_name
         self.gene_locations = []
-        self.genomic_location = genomic_location
+        self.genomic_position = genomic_location
 
     def get_gene_coordinates(self, gene:GeneAnnotationModel) -> None:
         """ Get the coordinates of the gene """
@@ -76,8 +76,7 @@ class CircRNAModel():
         circ_id = self.id
         tx_id = self.transcript_id
         gene_name = self.gene_name
-        chrom = self.fragments[0].chrom
         info = f'OFFSET={offset};LENGTH={length};INTRON={intron};' +\
             f'TRANSCRIPT_ID={tx_id};GENE_SYMBOL={gene_name};' +\
-            f'GENOMIC_POSITION={self.genomic_location}'
+            f'GENOMIC_POSITION={self.genomic_position}'
         return '\t'.join([gene_id, start, circ_id, '.', '.', '.', '.', info])